ref: move `parse_compressed_state_diffs` to common
