+++ state-reconstruct-fetcher/src/types/common.rs
@@ -1,7 +1,7 @@
 //! A collection of functions that get reused throughout format versions.
 use ethers::{abi, types::U256};
 
-use super::{PackingType, ParseError};
+use super::{L2ToL1Pubdata, PackingType, ParseError};
 use crate::constants::zksync::{LENGTH_BITS_OFFSET, OPERATION_BITMASK};
 
 pub struct ExtractedToken {
@@ -85,6 +85,67 @@ impl TryFrom<&abi::Token> for ExtractedToken {
 }
 
 // TODO: Move these to a dedicated parser struct.
+pub fn parse_compressed_state_diffs(
+    bytes: &[u8],
+    pointer: &mut usize,
+    end_bytes_to_ignore: usize,
+) -> Result<Vec<L2ToL1Pubdata>, ParseError> {
+    let mut state_diffs = Vec::new();
+    // Parse the header.
+    let version = u8::from_be_bytes(read_next_n_bytes(bytes, pointer));
+    if version != 1 {
+        return Err(ParseError::InvalidCompressedValue(String::from(
+            "VersionMismatch",
+        )));
+    }
+
+    if *pointer >= bytes.len() {
+        return Ok(state_diffs);
+    }
+
+    let mut buffer = [0; 4];
+    buffer[1..].copy_from_slice(&bytes[*pointer..*pointer + 3]);
+    *pointer += 3;
+    let _total_compressed_len = u32::from_be_bytes(buffer);
+
+    let enumeration_index = u8::from_be_bytes(read_next_n_bytes(bytes, pointer));
+
+    // Parse initial writes.
+    let num_of_initial_writes = u16::from_be_bytes(read_next_n_bytes(bytes, pointer));
+    for _ in 0..num_of_initial_writes {
+        let derived_key = U256::from_big_endian(&read_next_n_bytes::<32>(bytes, pointer));
+
+        let packing_type = read_compressed_value(bytes, pointer)?;
+        state_diffs.push(L2ToL1Pubdata::CompressedStateDiff {
+            is_repeated_write: false,
+            derived_key,
+            packing_type,
+        });
+    }
+
+    // Parse repeated writes.
+    while *pointer < bytes.len() - end_bytes_to_ignore {
+        let derived_key = match enumeration_index {
+            4 => U256::from_big_endian(&read_next_n_bytes::<4>(bytes, pointer)),
+            5 => U256::from_big_endian(&read_next_n_bytes::<5>(bytes, pointer)),
+            _ => {
+                return Err(ParseError::InvalidCompressedValue(String::from(
+                    "RepeatedDerivedKey",
+                )))
+            }
+        };
+
+        let packing_type = read_compressed_value(bytes, pointer)?;
+        state_diffs.push(L2ToL1Pubdata::CompressedStateDiff {
+            is_repeated_write: true,
+            derived_key,
+            packing_type,
+        });
+    }
+
+    Ok(state_diffs)
+}
+
 pub fn read_compressed_value(bytes: &[u8], pointer: &mut usize) -> Result<PackingType, ParseError> {
     let metadata = u8::from_be_bytes(read_next_n_bytes(bytes, pointer));
     let operation = metadata & OPERATION_BITMASK;
